layout: dark mode toggle button moved into menu drawer
